+++ src/main/java/seedu/address/logic/commands/listingcommands/ListingCommandsUtil.java
@@ -13,7 +13,7 @@ import seedu.address.model.listing.Area;
 import seedu.address.model.listing.Listing;
 import seedu.address.model.listing.Price;
 import seedu.address.model.listing.Region;
-import seedu.address.model.person.Name;
+import seedu.address.model.name.Name;
 import seedu.address.model.person.Person;
 import seedu.address.model.person.Role;
