+++ src/main/java/seedu/address/model/name/Name.java
@@ -1,10 +1,10 @@
-package seedu.address.model.person;
+package seedu.address.model.name;
 
 import static java.util.Objects.requireNonNull;
 import static seedu.address.commons.util.AppUtil.checkArgument;
 
 /**
- * Represents a Person's name in the address book.
+ * Represents a Person's name in the address book or a Listing's name in Listings.
  * Guarantees: immutable; is valid as declared in {@link #isValidName(String)}
  */
 public class Name {
