+++ src/main/java/seedu/address/model/person/Seller.java
@@ -3,6 +3,7 @@ package seedu.address.model.person;
 import java.util.Set;
 
 import seedu.address.model.appointment.Appointment;
+import seedu.address.model.name.Name;
 import seedu.address.model.tag.Tag;
 
 /**
