+++ src/test/java/seedu/address/logic/commands/clientcommands/appointmentcommands/AddAppointmentCommandTest.java
@@ -26,7 +26,7 @@ import seedu.address.model.appointment.Appointment;
 import seedu.address.model.appointment.Date;
 import seedu.address.model.appointment.From;
 import seedu.address.model.appointment.To;
-import seedu.address.model.person.Name;
+import seedu.address.model.name.Name;
 import seedu.address.model.person.Person;
 import seedu.address.testutil.PersonBuilder;
