+++ src/test/java/seedu/address/logic/commands/clientcommands/appointmentcommands/DeleteAppointmentCommandTest.java
@@ -28,8 +28,8 @@ import seedu.address.model.Model;
 import seedu.address.model.ModelManager;
 import seedu.address.model.UserPrefs;
 import seedu.address.model.appointment.Appointment;
+import seedu.address.model.name.Name;
 import seedu.address.model.person.Buyer;
-import seedu.address.model.person.Name;
 import seedu.address.model.person.Person;
 
 public class DeleteAppointmentCommandTest {
