+++ src/test/java/seedu/address/logic/commands/listingcommands/AddBuyersToListingCommandTest.java
@@ -36,7 +36,7 @@ import seedu.address.model.Model;
 import seedu.address.model.ModelManager;
 import seedu.address.model.UserPrefs;
 import seedu.address.model.listing.Listing;
-import seedu.address.model.person.Name;
+import seedu.address.model.name.Name;
 import seedu.address.model.person.Person;
 import seedu.address.testutil.ListingBuilder;
