+++ src/test/java/seedu/address/logic/commands/listingcommands/AddListingCommandTest.java
@@ -36,7 +36,7 @@ import seedu.address.model.ReadOnlyListings;
 import seedu.address.model.listing.Area;
 import seedu.address.model.listing.Listing;
 import seedu.address.model.listing.Region;
-import seedu.address.model.person.Name;
+import seedu.address.model.name.Name;
 import seedu.address.model.person.Person;
 import seedu.address.testutil.ListingBuilder;
