+++ src/test/java/seedu/address/logic/commands/listingcommands/ClearListingCommandTest.java
@@ -17,7 +17,7 @@ import seedu.address.model.listing.Area;
 import seedu.address.model.listing.Listing;
 import seedu.address.model.listing.Price;
 import seedu.address.model.listing.Region;
-import seedu.address.model.person.Name;
+import seedu.address.model.name.Name;
 import seedu.address.model.person.Person;
 import seedu.address.testutil.PersonBuilder;
