+++ src/test/java/seedu/address/logic/commands/listingcommands/RemoveBuyersFromListingCommandTest.java
@@ -34,7 +34,7 @@ import seedu.address.model.Model;
 import seedu.address.model.ModelManager;
 import seedu.address.model.UserPrefs;
 import seedu.address.model.listing.Listing;
-import seedu.address.model.person.Name;
+import seedu.address.model.name.Name;
 import seedu.address.testutil.ListingBuilder;
 
 public class RemoveBuyersFromListingCommandTest {
