+++ src/test/java/seedu/address/logic/parser/clientcommandparsers/AddClientProfileParserTest.java
@@ -27,9 +27,9 @@ import org.junit.jupiter.api.Test;
 import seedu.address.logic.Messages;
 import seedu.address.logic.commands.clientcommands.AddBuyerProfile;
 import seedu.address.logic.commands.clientcommands.AddSellerProfile;
+import seedu.address.model.name.Name;
 import seedu.address.model.person.Buyer;
 import seedu.address.model.person.Email;
-import seedu.address.model.person.Name;
 import seedu.address.model.person.Phone;
 import seedu.address.model.person.Seller;
 import seedu.address.testutil.PersonBuilder;
