+++ src/test/java/seedu/address/logic/parser/clientcommandparsers/EditClientCommandParserTest.java
@@ -28,7 +28,7 @@ import org.junit.jupiter.api.Test;
 import seedu.address.logic.Messages;
 import seedu.address.logic.commands.clientcommands.EditClientCommand;
 import seedu.address.logic.commands.clientcommands.EditClientCommand.EditPersonDescriptor;
-import seedu.address.model.person.Name;
+import seedu.address.model.name.Name;
 import seedu.address.model.tag.Tag;
 import seedu.address.testutil.EditPersonDescriptorBuilder;
