+++ src/test/java/seedu/address/logic/parser/clientcommandparsers/appointmentcommandparsers/AddAppointmentCommandParserTest.java
@@ -17,7 +17,7 @@ import seedu.address.model.appointment.Appointment;
 import seedu.address.model.appointment.Date;
 import seedu.address.model.appointment.From;
 import seedu.address.model.appointment.To;
-import seedu.address.model.person.Name;
+import seedu.address.model.name.Name;
 
 public class AddAppointmentCommandParserTest {
     private static final String VALID_DATE = "01-11-24";
