+++ src/test/java/seedu/address/logic/parser/listingcommandparsers/AddListingCommandParserTest.java
@@ -31,7 +31,7 @@ import seedu.address.model.listing.Address;
 import seedu.address.model.listing.Area;
 import seedu.address.model.listing.Price;
 import seedu.address.model.listing.Region;
-import seedu.address.model.person.Name;
+import seedu.address.model.name.Name;
 
 public class AddListingCommandParserTest {
