+++ src/test/java/seedu/address/model/person/SellerTest.java
@@ -11,6 +11,7 @@ import seedu.address.model.appointment.Appointment;
 import seedu.address.model.appointment.Date;
 import seedu.address.model.appointment.From;
 import seedu.address.model.appointment.To;
+import seedu.address.model.name.Name;
 import seedu.address.model.tag.Tag;
 
 public class SellerTest {
