+++ src/test/java/seedu/address/storage/JsonAdaptedListingTest.java
@@ -15,7 +15,7 @@ import seedu.address.model.listing.Address;
 import seedu.address.model.listing.Area;
 import seedu.address.model.listing.Price;
 import seedu.address.model.listing.Region;
-import seedu.address.model.person.Name;
+import seedu.address.model.name.Name;
 import seedu.address.model.person.Person;
 
 public class JsonAdaptedListingTest {
